onBeforeSave event to manage relationships
Added listener onBefore save to manage the relationships when an entry is saved.

Many to Many will reference its settings and see if the currently saved entry is in one of the set settings, it will then look for the associated field and sanitize the entries to only entries from within that section, it will then manage the relationship from both sides of the coin.